Fix updateCourses

+++ src/domain/repositories/coursesRepository.js
@@ -370,16 +370,19 @@ class coursesRepository {
       }
 
       if (titulo && titulo !== course.titulo) {
-        const newCourseTitle = titulo.toLowerCase().replace(/\s/g, "_");
-        const oldCourseTitle = course.titulo.toLowerCase().replace(/\s/g, "_");
-        const imgExtension = path.extname(course.url_portada);
+        if (course.url_portada !== null) {
+          const newCourseTitle = titulo.toLowerCase().replace(/\s/g, "_");
+          const oldCourseTitle = course.titulo.toLowerCase().replace(/\s/g, "_");
+          const imgExtension = path.extname(course.url_portada);
+  
+          const oldImagePath = `src/infraestructure/storage/local/cursos/${oldCourseTitle}${imgExtension}`;
+          const newImagePath = `src/infraestructure/storage/local/cursos/${newCourseTitle}${imgExtension}`;
+          fs.renameSync(oldImagePath, newImagePath);
 
-        const oldImagePath = `src/infraestructure/storage/local/cursos/${oldCourseTitle}${imgExtension}`;
-        const newImagePath = `src/infraestructure/storage/local/cursos/${newCourseTitle}${imgExtension}`;
-        fs.renameSync(oldImagePath, newImagePath);
+          course.url_portada = `${newCourseTitle}${imgExtension}`;
+        }
 
         course.titulo = titulo;
-        course.url_portada = `${newCourseTitle}${imgExtension}`;
       }
       course.url_video_intro = url_video_intro;
       course.descripcion = descripcion;
@@ -400,15 +403,19 @@ class coursesRepository {
         where: { idcurso, estado: 1 },
       });
 
-      const oldImageExtension = path.extname(course.url_portada);
-      const newImageExtension = path.extname(newImage);
+      if (course.url_portada == null) {
+        course.url_portada = newImage;
+      } else {
+        const oldImageExtension = path.extname(course.url_portada);
+        const newImageExtension = path.extname(newImage);
+  
+        if (newImageExtension !== oldImageExtension) {
+          const oldImagePath = `src/infraestructure/storage/local/cursos/${course.url_portada}`;
+          fs.unlinkSync(oldImagePath);
 
-      if (newImageExtension !== oldImageExtension) {
-        const oldImagePath = `src/infraestructure/storage/local/cursos/${course.url_portada}`;
-        fs.unlinkSync(oldImagePath);
+          course.url_portada = newImage;
+        }
       }
-
-      course.url_portada = newImage;
 
       await course.save();
       return course;
